# services/gateway/app/routers/tenants.py
"""Tenant admin endpoints used by the admin dashboard."""

from __future__ import annotations

import uuid
from datetime import datetime
from typing import Any

from fastapi import APIRouter, Depends, HTTPException, Query, status

from packages.db import get_db
from packages.models import Tenant, TenantAPIKey
from packages.schemas import (
    ErrorResponse,
    TenantAPIKeyCreateResponse,
    TenantAPIKeyResponse,
    TenantCreate,
    TenantResponse,
    TenantUpdate,
)
from ..auth.api_keys import issue_api_key, list_tenant_api_keys

router = APIRouter(prefix="/v1/tenants", tags=["tenants"])

def _tenant_to_response(tenant: Tenant) -> TenantResponse:
    return TenantResponse(
        tenant_id=tenant.id,
        name=tenant.name,
        plan=getattr(tenant, "plan", "free"),
        is_active=tenant.is_active,
        enable_premium_escalation=getattr(tenant, "enable_premium_escalation", False),
        enable_semantic_cache=getattr(tenant, "enable_semantic_cache", False),
        cache_similarity_threshold=getattr(tenant, "cache_similarity_threshold", 0.92),
        max_requests_per_day=getattr(tenant, "max_requests_per_day", None),
        created_at=tenant.created_at,
        updated_at=getattr(tenant, "updated_at", tenant.created_at),
    )


def _get_record(db: Any, model: type[Any], record_id: str) -> Any | None:
    if hasattr(db, "get"):
        return db.get(model, record_id)
    if hasattr(db, "query"):
        return db.query(model).filter_by(id=record_id).first()
    raise HTTPException(status_code=500, detail="Database session does not support record lookup.")


def _list_records(db: Any, model: type[Any]) -> list[Any]:
    if hasattr(db, "_objects"):
        return [record for record in db._objects.values() if isinstance(record, model)]
    if hasattr(db, "query"):
        return list(db.query(model).all())
    raise HTTPException(status_code=500, detail="Database session does not support record listing.")


def create_tenant_record(request: TenantCreate, db: object) -> TenantResponse:
    now = datetime.utcnow()
    tenant = Tenant(
        id=f"tenant_{uuid.uuid4().hex}",
        name=request.name.strip(),
        is_active=request.is_active,
        created_at=now,
    )
    tenant.updated_at = now
    tenant.plan = getattr(request, "plan", "free")
    tenant.enable_premium_escalation = getattr(request, "enable_premium_escalation", False)
    tenant.enable_semantic_cache = getattr(request, "enable_semantic_cache", False)
    tenant.cache_similarity_threshold = 0.92
    tenant.max_requests_per_day = None
    db.add(tenant)
    if hasattr(db, "commit"):
        db.commit()
    return _tenant_to_response(tenant)


def _api_key_to_response(record: TenantAPIKey) -> TenantAPIKeyResponse:
    return TenantAPIKeyResponse(
        api_key_id=record.id,
        tenant_id=record.tenant_id,
        key_prefix=record.key_prefix,
        is_active=record.is_active,
        created_at=record.created_at,
        updated_at=record.updated_at,
        last_used_at=record.last_used_at,
    )


@router.get(
    "",
    response_model=list[TenantResponse],
    summary="List tenants",
    description="Returns tenant records, optionally filtered to active tenants only.",
)
def list_tenants(
    active_only: bool = Query(default=False),
    db: object = Depends(get_db),
) -> list[TenantResponse]:
    tenants = [
        tenant
        for tenant in _list_records(db, Tenant)
        if isinstance(tenant, Tenant) and (tenant.is_active or not active_only)
    ]
    tenants.sort(key=lambda tenant: tenant.created_at, reverse=True)
    return [_tenant_to_response(tenant) for tenant in tenants]


@router.post(
    "",
    response_model=TenantResponse,
    summary="Create tenant",
    description="Creates a tenant record that can later issue tenant-scoped API keys.",
    responses={
        status.HTTP_400_BAD_REQUEST: {
            "model": ErrorResponse,
            "description": "Tenant payload is invalid.",
        }
    },
)
def create_tenant(
    request: TenantCreate,
    db: object = Depends(get_db),
) -> TenantResponse:
    return create_tenant_record(request, db)


@router.get(
    "/{tenant_id}",
    response_model=TenantResponse,
    summary="Get tenant",
    responses={
        status.HTTP_404_NOT_FOUND: {
            "model": ErrorResponse,
            "description": "Tenant was not found.",
        }
    },
)
def get_tenant(tenant_id: str, db: object = Depends(get_db)) -> TenantResponse:
    tenant = _get_record(db, Tenant, tenant_id)
    if tenant is None or not isinstance(tenant, Tenant):
        raise HTTPException(status_code=404, detail=f"Tenant {tenant_id} not found")
    return _tenant_to_response(tenant)


@router.patch(
    "/{tenant_id}",
    response_model=TenantResponse,
    summary="Update tenant",
    responses={
        status.HTTP_404_NOT_FOUND: {
            "model": ErrorResponse,
            "description": "Tenant was not found.",
        }
    },
)
def update_tenant(
    tenant_id: str,
    request: TenantUpdate,
    db: object = Depends(get_db),
) -> TenantResponse:
    tenant = _get_record(db, Tenant, tenant_id)
    if tenant is None or not isinstance(tenant, Tenant):
        raise HTTPException(status_code=404, detail=f"Tenant {tenant_id} not found")

    updates = request.model_dump(exclude_unset=True)

    for field, value in updates.items():
        setattr(tenant, field, value)

    tenant.updated_at = datetime.utcnow()

    if hasattr(db, "commit"):
        db.commit()
    return _tenant_to_response(tenant)


@router.get(
    "/{tenant_id}/api-keys",
    response_model=list[TenantAPIKeyResponse],
    summary="List tenant API keys",
    responses={
        status.HTTP_404_NOT_FOUND: {
            "model": ErrorResponse,
            "description": "Tenant was not found.",
        }
    },
)
def list_api_keys(tenant_id: str, db: object = Depends(get_db)) -> list[TenantAPIKeyResponse]:
    tenant = _get_record(db, Tenant, tenant_id)
    if tenant is None or not isinstance(tenant, Tenant):
        raise HTTPException(status_code=404, detail=f"Tenant {tenant_id} not found")
    return [_api_key_to_response(record) for record in list_tenant_api_keys(db, tenant_id)]


@router.post(
    "/{tenant_id}/api-keys",
    response_model=TenantAPIKeyCreateResponse,
    summary="Create tenant API key",
    description="Issues a new plaintext API key and returns it once at creation time.",
    responses={
        status.HTTP_404_NOT_FOUND: {
            "model": ErrorResponse,
            "description": "Tenant was not found.",
        }
    },
)
def create_api_key(tenant_id: str, db: object = Depends(get_db)) -> TenantAPIKeyCreateResponse:
    tenant = _get_record(db, Tenant, tenant_id)
    if tenant is None or not isinstance(tenant, Tenant):
        raise HTTPException(status_code=404, detail=f"Tenant {tenant_id} not found")

    issued = issue_api_key(tenant_id=tenant_id)
    db.add(issued.record)
    if hasattr(db, "commit"):
        db.commit()

    return TenantAPIKeyCreateResponse(
        api_key_id=issued.record.id,
        tenant_id=issued.record.tenant_id,
        key_prefix=issued.record.key_prefix,
        api_key=issued.plaintext,
        created_at=issued.record.created_at,
    )